requirements.txt
# ============================================================
# StudyHub – Merged Production Requirements
# Fused from: requirements-3.txt + requirements-5.txt
# Python: 3.11.x  (recommended for Render deployment)
# ============================================================


# ── Core Web Framework ────────────────────────────────────
Flask==3.0.3
Werkzeug==3.0.4                 # Must match Flask 3.0.x series


# ── Database ──────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Migrate==4.0.7
alembic==1.13.1
psycopg2-binary==2.9.9          # PostgreSQL driver (Supabase / DATABASE_URL)


# ── Authentication & Security ─────────────────────────────
Flask-Login==0.6.3              # UserMixin, login_required
PyJWT==2.8.0                    # jwt.encode/decode in auth.py, utils.py
bleach==6.1.0                   # bleach.clean() in websocket_messages.py


# ── Email ─────────────────────────────────────────────────
Flask-Mail==0.10.0              # Mail(), Message in extensions.py / utils.py


# ── Real-time / WebSocket ─────────────────────────────────
Flask-SocketIO==5.3.6
python-socketio==5.11.4         # Required peer dep for Flask-SocketIO
python-engineio==4.9.1          # Required by python-socketio
eventlet==0.35.2                # Async worker for Flask-SocketIO
bidict==0.23.1                  # Required internally by python-socketio


# ── Google OAuth ──────────────────────────────────────────
Flask-Dance==7.0.1              # make_google_blueprint in auth.py
oauthlib==3.2.2                 # Required by Flask-Dance
requests-oauthlib==1.3.1        # Required by Flask-Dance


# ── Rate Limiting ─────────────────────────────────────────
Flask-Limiter==3.5.0
limits==3.7.0                   # Required by Flask-Limiter


# ── Task Scheduling ───────────────────────────────────────
APScheduler==3.10.4             # Background job scheduler


# ── Firebase Push Notifications ───────────────────────────
firebase-admin==6.5.0           # PushNotificationService


# ── File Storage / Media ──────────────────────────────────
cloudinary==1.38.0              # Posts, messages, storage uploads


# ── HTTP Client ───────────────────────────────────────────
requests==2.31.0                # OAuth + Supabase uploads


# ── Environment / Config ──────────────────────────────────
python-dotenv==1.0.0            # Loads .env variables


# ── Production WSGI Server ────────────────────────────────
gunicorn==21.2.0                # Render start command: gunicorn app:app --worker-class eventlet


# ── Utilities / Internal Deps ─────────────────────────────
greenlet==3.0.3                 # Required by SQLAlchemy async + eventlet
blinker==1.8.2                  # Required by Flask signals and Flask-Dance
